Refine watch contract DTO placement (#4169)
